+++ earth2studio/data/__init__.py
@@ -29,6 +29,7 @@ from .goes import GOES
 from .hrrr import HRRR, HRRR_FX
 from .isd import ISD
 from .jpss import JPSS
+from .jpss_atms import JPSS_ATMS
 from .metop_amsua import MetOpAMSUA
 from .metop_avhrr import MetOpAVHRR
 from .mrms import MRMS